+++ effective-gear/fabric/src/main/java/com/chimericdream/effectivegear/fabric/worldgen/EffectiveGearTrimProvider.java
@@ -1,0 +1,25 @@
+package com.chimericdream.effectivegear.fabric.worldgen;
+
+import net.fabricmc.fabric.api.datagen.v1.FabricPackOutput;
+import net.fabricmc.fabric.api.datagen.v1.provider.FabricDynamicRegistryProvider;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.core.registries.Registries;
+import org.jspecify.annotations.NonNull;
+
+import java.util.concurrent.CompletableFuture;
+
+public class EffectiveGearTrimProvider extends FabricDynamicRegistryProvider {
+    public EffectiveGearTrimProvider(FabricPackOutput output, CompletableFuture<HolderLookup.Provider> registriesFuture) {
+        super(output, registriesFuture);
+    }
+
+    @Override
+    protected void configure(HolderLookup.Provider registries, Entries entries) {
+        entries.addAll(registries.lookupOrThrow(Registries.TRIM_MATERIAL));
+    }
+
+    @Override
+    public @NonNull String getName() {
+        return "Effective Gear Trim Materials";
+    }
+}